Reorganize protocol generators
